move fauxbmotion.py from bapsflib._hdf.maps.controls.tests -> bapsflib.faux.controls
