+++ bapsflib/_hdf/maps/controls/tests/__init__.py
@@ -8,7 +8,6 @@
 # License: Standard 3-clause BSD; see "LICENSES/LICENSE.txt" for full
 #   license terms and contributor agreement.
 #
-from bapsflib._hdf.maps.controls.tests.fauxbmotion import FauxBMotion
 from bapsflib._hdf.maps.controls.tests.fauxn5700ps import FauxN5700PS
 from bapsflib._hdf.maps.controls.tests.fauxnixyz import FauxNIXYZ
 from bapsflib._hdf.maps.controls.tests.fauxnixz import FauxNIXZ
