+++ bapsflib/faux/controls/fauxbmotion.py
